Consolidating to single mock user for test consistency
