+++ src/applications/combined-debt-portal/debt-letters/tests/e2e/debt-summary-card-content.cypress.spec.js
@@ -1,5 +1,5 @@
 import mockFeatureToggles from '../../../combined/tests/e2e/fixtures/mocks/feature-toggles.json';
-import mockUser from './fixtures/mocks/mock-user.json';
+import mockUser from '../../../combined/tests/e2e/fixtures/mocks/mock-user-81.json';
 import {
   copayResponses,
   debtResponses,
