+++ src/applications/combined-debt-portal/medical-copays/tests/e2e/copay-statement-content.cypress.spec.js
@@ -6,7 +6,7 @@
  * @testrailinfo runName MCP-e2e-Statements
  */
 import mockFeatureToggles from '../../../combined/tests/e2e/fixtures/mocks/feature-toggles.json';
-import mockUser from './fixtures/mocks/mock-user.json';
+import mockUser from '../../../combined/tests/e2e/fixtures/mocks/mock-user-81.json';
 import {
   copayResponses,
   debtResponses,
