--- src/applications/combined-debt-portal/medical-copays/tests/e2e/fixtures/mocks/mock-user.json
@@ -1,97 +1,0 @@
-{
-  "data": {
-    "id": "",
-    "type": "users_scaffolds",
-    "attributes": {
-      "services": [
-        "facilities",
-        "hca",
-        "edu-benefits",
-        "form-save-in-progress",
-        "form-prefill",
-        "evss-claims",
-        "add-person",
-        "user-profile",
-        "appeals-status",
-        "identity-proofed"
-      ],
-      "account": {
-        "accountUuid": "278bb426-e69b-49ec-a7e8-a1b8db3ddd44"
-      },
-      "profile": {
-        "email": "vets.gov.user+81@gmail.com",
-        "firstName": "TRAVIS",
-        "middleName": "D",
-        "lastName": "JONES",
-        "birthDate": "1950-09-06",
-        "gender": "F",
-        "zip": "44102",
-        "lastSignedIn": "2021-11-30T15:31:09.659Z",
-        "loa": {
-          "current": 3,
-          "highest": 3
-        },
-        "multifactor": true,
-        "verified": true,
-        "signIn": {
-          "serviceName": "idme",
-          "accountType": "N/A"
-        },
-        "authnContext": "http://idmanagement.gov/ns/assurance/loa/3"
-      },
-      "vaProfile": {
-        "status": "OK",
-        "birthDate": "19510710",
-        "familyName": "Jones",
-        "gender": "M",
-        "givenNames": ["Travis", "Null"],
-        "isCernerPatient": false,
-        "facilities": [],
-        "vaPatient": false,
-        "mhvAccountState": "OK"
-      },
-      "veteranStatus": {
-        "status": "OK",
-        "isVeteran": false,
-        "servedInMilitary": false
-      },
-      "inProgressForms": [],
-      "prefillsAvailable": [
-        "21-686C",
-        "40-10007",
-        "0873",
-        "22-1990",
-        "22-1990N",
-        "22-1990E",
-        "22-1995",
-        "22-5490",
-        "22-5495",
-        "22-0993",
-        "22-0994",
-        "FEEDBACK-TOOL",
-        "22-10203",
-        "22-1990S",
-        "22-1990EZ",
-        "21-526EZ",
-        "1010ez",
-        "21P-530",
-        "21P-527EZ",
-        "686C-674",
-        "20-0996",
-        "10182",
-        "MDOT",
-        "5655",
-        "28-8832",
-        "28-1900"
-      ],
-      "vet360ContactInformation": {},
-      "session": {
-        "ssoe": true,
-        "transactionid": "c0CeCHB1m4PO50ZwxesueRn/8MbWaLgIOrTQSuB/Fbo="
-      }
-    }
-  },
-  "meta": {
-    "errors": null
-  }
-}
